+++ CheckDrive.Api/CheckDrive.Domain/Entities/CarUsageSummary.cs
@@ -2,8 +2,8 @@
 
 public class CarUsageSummary
 {
-    public int CurrentMonthDistance { get; set; }
-    public int CurrentYearDistance { get; set; }
+    public decimal CurrentMonthDistance { get; set; }
+    public decimal CurrentYearDistance { get; set; }
     public decimal CurrentMonthFuelConsumption { get; set; }
     public decimal CurrentYearFuelConsumption { get; set; }
 }
